requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
azureml-defaults==1.48.0
azureml-core==1.48.0
inference-schema==1.5.1
azureml-mlflow==1.48.0
azureml-contrib-services==1.48.0